Feature/loginscreen

+++ pages/api/projects/[id].js
@@ -1,7 +1,11 @@
 import dbConnect from "@/db/connect";
 import Project from "@/db/models/ProjectSchema";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(request, response) {
+  const session = await getServerSession(request, response, authOptions);
+
   await dbConnect();
   const { id } = request.query;
 
@@ -15,13 +19,13 @@ export default async function handler(request, response) {
     response.status(200).json(project);
   }
 
-  if (request.method === "PUT") {
+  if (session && request.method === "PUT") {
     const updatedProject = request.body;
     await Project.findByIdAndUpdate(id, updatedProject);
     response.status(200).json({ status: `Project successfully updated!` });
   }
 
-  if (request.method === "DELETE") {
+  if (session && request.method === "DELETE") {
     const deletedProject = await Project.findByIdAndDelete(id);
     if (deletedProject) {
       response.status(200).json({ status: `Project successfully deleted.` });
